requirements.txt
streamlit>=1.21.0
altair<5
pytest

google-generativeai
python-dotenv